issuer validation handling in `JWSHelper` and split JWSHelper into separate files
