+++ pyeudiw/openid4vp/direct_post_response.py
@@ -2,8 +2,10 @@ import logging
 
 from typing import Dict
 from pyeudiw.jwk import JWK
-from pyeudiw.jwt import JWEHelper, JWSHelper
+
 from pyeudiw.jwk.exceptions import KidNotFoundError
+from pyeudiw.jwt.jwe_helper import JWEHelper
+from pyeudiw.jwt.jws_helper import JWSHelper
 from pyeudiw.jwt.utils import decode_jwt_header, is_jwe_format
 from pyeudiw.openid4vp.exceptions import (
     InvalidVPToken,
