+++ pyeudiw/tests/jwt/test_sign_verify.py
@@ -1,5 +1,6 @@
 import pytest
-from pyeudiw.jwt import JWSHelper
+
+from pyeudiw.jwt.jws_helper import JWSHelper
 from pyeudiw.jwt.utils import decode_jwt_header
