+++ pyeudiw/tests/oauth2/test_dpop.py
@@ -3,7 +3,8 @@ import hashlib
 import pytest
 
 
-from pyeudiw.jwt import JWSHelper
+
+from pyeudiw.jwt.jws_helper import JWSHelper
 from pyeudiw.jwt.utils import decode_jwt_header, decode_jwt_payload
 from pyeudiw.oauth2.dpop import DPoPIssuer, DPoPVerifier
 from pyeudiw.oauth2.dpop.exceptions import InvalidDPoPKid
